# api/app/docs/swagger.yaml
definitions:
  dto.CreateReportDTO:
    properties:
      external_id:
        type: string
      image_url:
        type: string
    required:
    - external_id
    - image_url
    type: object
  models.Report:
    properties:
      created_at:
        type: string
      external_id:
        type: string
      id:
        type: string
      image_url:
        type: string
      priority:
        type: integer
      status:
        type: integer
    required:
    - external_id
    - image_url
    type: object
host: localhost:3000
info:
  contact: {}
  description: A quickly API setup to evaulate avatar.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Image Evaluate API
  version: "1.0"
paths:
  /api/report/{id}:
    put:
      consumes:
      - application/json
      description: Update the report corresponding to the input id
      parameters:
      - description: ID of the report to be updated
        in: path
        name: id
        required: true
        type: string
      - description: Create report
        in: body
        name: report
        required: true
        schema:
          $ref: '#/definitions/models.Report'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Report'
      security:
      - ApiKeyAuth: []
      summary: Update report identified by the given id
      tags:
      - report
  /report:
    get:
      consumes:
      - application/json
      description: List all reports
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Report'
            type: array
      summary: Get a list of reports id
      tags:
      - report
    post:
      consumes:
      - application/json
      description: Saves a request for avatar evaluate
      parameters:
      - description: Create a report
        in: body
        name: evaluate
        required: true
        schema:
          $ref: '#/definitions/dto.CreateReportDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Report'
            type: array
      summary: Endpoint to save avatar reports
      tags:
      - report
  /report/{id}:
    get:
      consumes:
      - application/json
      description: Get details of report corresponding to the input id
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Report'
      summary: Get details for a given report id
      tags:
      - report
swagger: "2.0"